// src/api/generated/models/documentRelationshipsNodesItemDocumentType.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */

export type DocumentRelationshipsNodesItemDocumentType = typeof DocumentRelationshipsNodesItemDocumentType[keyof typeof DocumentRelationshipsNodesItemDocumentType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const DocumentRelationshipsNodesItemDocumentType = {
  MATERIAL_REQUEST: 'MATERIAL_REQUEST',
  SUPPLIER_QUOTE: 'SUPPLIER_QUOTE',
  INVOICE: 'INVOICE',
  PAYMENT: 'PAYMENT',
  SHIPPING_INVOICE: 'SHIPPING_INVOICE',
  CONSOLIDATION: 'CONSOLIDATION',
} as const;